made inputsample a float. Switched order of m_Rate and m_Depth

+++ CSD2c/Prototype/datorro/modallpass.cpp
@@ -18,7 +18,7 @@ ModAllPass::~ModAllPass() {
     circ = nullptr;
 }
 
-float ModAllPass::output(double inputSample) {
+float ModAllPass::output(float inputSample) {
     circ->writeToBuffer(inputSample + (tempSample * m_coeffG));
     tempSample = circ->readFromBuffer();
     circ->incrementIndeces();
